Modified the name of the AllJoyn_DaemonService and AllJoyn_DaemonUser
While working on the documentation for the most recent web release we chose to name the Android Service AllJoyn_DaemonService for the service, and AllJoyn_DaemonUser for the launching application.  I have changed the names of the projects to reflect that name change.  I have added the .project file for the AllJoyn_DaemonUser project so we can just import the project into eclipse. I have added a .gitignore file for the AllJoyn_DeamonUser so that it does not report the gen folder and the bin folder as new items.