ALLJOYN-391 Fixed posible NULL pointer dereference.
If was possible for the replyCode to be set to ALLJOYN_JOINSESSION_REPLY_NO_SESSION or ALLJOYN_JOINSESSION_REPLY_UNREACHABLE and have the b2bEp pointer not be initilised.  This would result in the code trying to dereference a NULL pointer when making the b2bEp->DecrementRef() call. This check will fix this potintal problem.