ALLJOYN-391 removal of un-needed NULL check
The code is checking g_msgBus to see if it is NULL before calling delete on it.  Since it was allocated with the 'new' operator which throws std::bad_alloc when it can not allocate memor.  If this point is reached in the code we know that g_msgBus was allocated.