Addition of Doxygen config file that will run doxygen on all header files, even files not found in the public API .
If scons is run with DOCS=dev doxygen will run on all header files and will include all docuementation that is marked with the @internal tag as well as code that is marked @cond ALLJOYN_DEV ... @endcond.  Since doxygen has not been run on many of the files using this command will currently result in a lot of warnings from doxygen.  This is intended to help improve the documentation of all files not only the public API.