Modified linux chat sample to build and run in both linux and windows
Removed the call to 'getline(...)' which is a function call that is not in the ANSI C standard. Replaced it with new function 'get_line(...)'  that uses 'fgets' to read user input. The new get_line function does not allocate memory like getline() so it will ignore any user input that does not fit in the pre allocated char array.

Uses a different connetSpec value for windows since unix:abstract=alljoyn will not work in windows.

Calls 'Sleep' in place of 'sleep' when running in windows.