ALLJOYN-463 fixed doxygen warnings.
Doxygen was skipping things marked with the keyword 'static' since this is often used for information hidding in C code it made since to skip this infomation however our code is not using the keywork 'static' for information hidding and we expect users to know about static constants.  I changed the doxygen config files so things marked with 'static' are documented.  Fixed other misc. documentation issues at the same time.