Moved UpdateDelegations() to run on the BusDispatcher thread.
Updating local advertisements the BlueZ implementation of BTAccessor involved
method calls to a different instance of BusAttachement, but if the thread that
UpdateDelegations() was running on was stopping, those method calls would be
locally errored out.  Moving UpdateDelegations() to run on the BusDispatcher
thread ensures that those advertisement updates complete.