Add BusAttachment::JoinSessionAsync() method.
This is an async version of BusAttachment::JoinSession().  It includes
the associated class BusAttachment::JoinSessionAsyncCB for the
callback.

Updated the documentation for ProxyBusObject::Listener to include a
comment that the preferred pattern for async public APIs is the one
used by BusAttachment::JoinSessionAsync(), not the one used for
ProxyBusObject::IntrospectRemoteObjectAsync().