[ALLJOYN-723]Prevent daemon sending out incorrect SessionLost signals.
When the name owner change signal was processed by the daemon it was sending out the SessionLost signal to all non-multipoint sessions not just the session that just ended.  This change will prevent the SessionLost signal being sent when it should not have been. This change should also address ALLJOYN-638 and ALLJOYN-727